+++ packages/search/src/types/SearchTargetPage.ts
@@ -1,0 +1,14 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+export interface SearchTargetPage {
+    numPages: number;
+    pageIndex: number;
+}
+
+export type SearchTargetPageFilter = (targetPage: SearchTargetPage) => boolean;